Lecture#25/requirements.txt
anyio==3.7.0
asgiref==3.6.0
certifi==2023.5.7
Django==4.2.1
django-dotenv==1.4.2
django-extensions==3.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
python-telegram-bot==20.3
pytz==2023.3
sniffio==1.3.0
sqlparse==0.4.4
